requirements.txt
# FlexCube AI Assistant - Python Dependencies
# Python 3.11+ required

# Core RAG Framework
llama-index>=0.10.0
llama-index-vector-stores-qdrant>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0

# Embeddings Model - flexible versions for compatibility
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.19.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP Client for Ollama
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Authentication & Security (Phase 7)
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
email-validator>=2.0.0

# Qdrant Client
qdrant-client>=1.7.0

# Document processing
docx2txt>=0.8
PyMuPDF>=1.23.0  # For PDF image extraction
Pillow>=10.0.0  # For image processing

# Browser Testing (E2E)
selenium>=4.15.0
webdriver-manager>=4.0.0
